Store decomposed, autocomplete to be fixed
