--- src/api/initial_config.js
@@ -1,6 +1,0 @@
-export default {
-  baseURL: 'http://localhost:8000/',
-  headers: {
-    accept: 'application/json'
-  }
-}
