+++ src/components/ProfileMenu.vue
@@ -23,7 +23,9 @@
 </template>
 
 <script>
-export default {}
+export default {
+  name: 'profile-menu'
+}
 </script>
 
 <style lang="scss">
