+++ src/store/actions.js
@@ -1,0 +1,8 @@
+export default {
+  SET_COOKIE({ commit }, payload) {
+    commit('SET_COOKIE', payload)
+  },
+  SET_USER_DATA({ commit }, payload) {
+    commit('SET_USER_DATA', payload)
+  }
+}
