+++ src/store/getters.js
@@ -1,0 +1,3 @@
+export default {
+  AUTH_STATUS: ({ isAuthorized }) => isAuthorized
+}
